Full implementation, finally. We delete old log records and update the lastPayoutRecalc property. Also, our early AVRG's are going to be simple, recalculate the factor every 20 payouts.

+++ src/java/com/threerings/msoy/avrg/server/persist/QuestLogSummaryRecord.java
@@ -42,5 +42,5 @@ public class QuestLogSummaryRecord extends PersistentRecord
     public int payoutCount;
     
     /** The sum of payout factors this AVRG has granted for this summary. */
-    public int payoutFactorTotal;
+    public float payoutFactorTotal;
 }
